requirements.txt
Django>=3.0.6,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0
flake8>=3.8.1,<3.9.0
coverage>=5.3.0,<5.4.0
django-rest-knox>=4.1.0,<4.2.0
beautifulsoup4>=4.9.0,<5.0.0